requirements.txt
transformers==4.44.2
datasets==2.19.1
hydra-core==1.3.2
ipython==8.12.0
lightning==2.2.5
numpy==1.24.0
omegaconf==2.3.0
pandas==2.2.2
peft==0.11.1
python-dotenv==1.0.1
PyYAML==6.0.2
scikit_learn==1.4.1.post1
timm==0.8.3.dev0
torch==2.2.2
tqdm==4.65.0
wandb==0.16.3
bitsandbytes==0.43.3
openpyxl==3.1.5